Minor Updates
1. Moved Clock assignment
2. Comments updated
3. Changed to three bytes returned to confirm this:

From http://i2c2p.twibright.com/spec/i2c.pdf page 10:

If a master-receiver is involved in a transfer, it must signal
the end of data to the slave- transmitter by not generating
an acknowledge on the last byte that was clocked out of
the slave. The slave-transmitter must release the data line
to allow the master to generate a STOP or repeated
START condition.